[archimedes]: Fixed minor issue

+++ server/src/experiments/archimedes.py
@@ -47,6 +47,8 @@ class Archimedes(Experiment):
 
         self._lock = threading.Lock()
 
+        self._workpool = None
+
         self._cfg_manager = cfg_manager
 
         # IP of the board, raspberry, beagle, or whatever.
@@ -287,7 +289,8 @@ class Archimedes(Experiment):
             print "[Archimedes] do_dispose called"
 
         # Finish the thread pool
-        self._workpool.terminate()
+        if self._workpool is not None:
+            self._workpool.terminate()
 
         return "ok"
